Use shard suffix in the security metadata id
The primary purpose here is to be able to distinguish between
different version of the same database name. We shouldn't have
security properties be inherited when a database is recreated. Adding
the suffix to the metadata id allows us to make this distinction, but
it also allows us to lazily deal with clearing out security docs for
old deleted database.